+++ QuickSort/quicksort.cpp
@@ -4,8 +4,8 @@
 std::ifstream fin("date.in");
 std::ofstream fout("date.out");
 
-long long partition(std::vector<double>& arr, long long low, long long high) {
-    double pivot = arr[high];  //pivot ca ultimul element
+long long partition(std::vector<long long>& arr, long long low, long long high) {
+    long long pivot = arr[high];  //pivot ca ultimul element
     long long i = low - 1;
 
     for (long long j = low; j < high; j++) {
@@ -20,7 +20,7 @@ long long partition(std::vector<double>& arr, long long low, long long high) {
     return i + 1;  // returnez unde fac partitionarea
 }
 
-void quicksort(std::vector<double>& arr, long long low, long long high) {
+void quicksort(std::vector<long long>& arr, long long low, long long high) {
     if (low < high) {
         long long part = partition(arr, low, high);
 
@@ -30,14 +30,14 @@ void quicksort(std::vector<double>& arr, long long low, long long high) {
 }
 
 int main() {
-    std::vector<double> arr;
-    double num;
+    std::vector<long long> arr;
+    long long num;
     while(fin >> num)
         arr.push_back(num);
 
     quicksort(arr, 0, arr.size() - 1);
 
-    for (double x : arr) {
+    for (long long x : arr) {
         fout << x << " ";
     }
