+++ QuickSort/quicksort_medianadin3.cpp
@@ -5,7 +5,7 @@
 std::ifstream fin("date.in");
 std::ofstream fout("date.out");
 
-long long medianadintrei(std::vector<double>& arr, long long low, long long high) {
+long long medianadintrei(std::vector<long long>& arr, long long low, long long high) {
     long long mid = low + (high - low) / 2;
 
     //gasesc mijlocul dintre cele 3 valori
@@ -17,9 +17,9 @@ long long medianadintrei(std::vector<double>& arr, long long low, long long high
         std::swap(arr[low], arr[mid]);
     return mid;
 }
-long long partition(std::vector<double>& arr, long long low, long long high) {
+long long partition(std::vector<long long>& arr, long long low, long long high) {
     long long pivotindex = medianadintrei(arr, low, high);
-    double pivot = arr[pivotindex];  //alegere pivot
+    long long pivot = arr[pivotindex];  //alegere pivot
     std::swap(arr[pivotindex], arr[high]);  //l am mutat ultimul ca altfel nush
     long long i = low - 1;
 
@@ -34,7 +34,7 @@ long long partition(std::vector<double>& arr, long long low, long long high) {
     return i + 1;  // returnez unde fac partitionarea
 }
 
-void quicksort(std::vector<double>& arr, long long low, long long high) {
+void quicksort(std::vector<long long>& arr, long long low, long long high) {
     if (low < high) {
         long long part = partition(arr, low, high);
 
@@ -44,8 +44,8 @@ void quicksort(std::vector<double>& arr, long long low, long long high) {
 }
 
 int main() {
-    std::vector<double> arr;
-    double num;
+    std::vector<long long> arr;
+    long long num;
     while(fin >> num)
         arr.push_back(num);
     quicksort(arr, 0, arr.size() - 1);
