+++ QuickSort/quicksort_randompivot.cpp
@@ -11,9 +11,9 @@ long long pivotrandom(long long a,long long b) {
     std::uniform_int_distribution<long long> dis(a, b);
     return dis(gen);
 }
-long long partition(std::vector<double>& arr, long long low, long long high) {
+long long partition(std::vector<long long>& arr, long long low, long long high) {
     long long pivotindex = pivotrandom(low, high);
-    double pivot = arr[pivotindex];  //alegere pivot
+    long long pivot = arr[pivotindex];  //alegere pivot
     std::swap(arr[pivotindex], arr[high]);  //l am mutat ultimul ca altfel nush
     long long i = low - 1;
 
@@ -29,7 +29,7 @@ long long partition(std::vector<double>& arr, long long low, long long high) {
     return i + 1;  // returnez unde fac partitionarea
 }
 
-void quicksort(std::vector<double>& arr, long long low, long long high) {
+void quicksort(std::vector<long long>& arr, long long low, long long high) {
     if (low < high) {
         long long part = partition(arr, low, high);
         
@@ -39,8 +39,8 @@ void quicksort(std::vector<double>& arr, long long low, long long high) {
 }
 
 int main() {
-    std::vector<double> arr;
-    double num;
+    std::vector<long long> arr;
+    long long num;
     while(fin >> num)
         arr.push_back(num);
